refactor: more cleanly separate view/model
