+++ src/story/pointer/external_event.rs
@@ -1,8 +1,9 @@
 use crate::story::{
     ant::commands::AntCommandsExt,
-    common::{position::Position, ui::{SelectedEntity, VisibleGrid}},
+    common::position::Position,
     crater_simulation::crater::Crater,
     grid::{Grid, VisibleGridState},
+    nest_rendering::common::{SelectedEntity, VisibleGrid},
     nest_simulation::nest::{AtNest, Nest},
 };
