+++ src/story/ui/action_menu.rs
@@ -4,8 +4,8 @@ use bevy::{prelude::*, window::PrimaryWindow};
 use bevy_egui::{egui, EguiContexts};
 
 use crate::settings::Settings;
-use crate::story::common::ui::VisibleGrid;
 use crate::story::crater_simulation::crater::Crater;
+use crate::story::nest_rendering::common::VisibleGrid;
 use crate::story::nest_simulation::nest::Nest;
 use crate::story::pointer::ExternalSimulationEvent;
 use crate::story::story_time::StoryTime;
